requirements_api.txt
fastapi
uvicorn[standard]
chromadb==1.0.13
telethon
sentence-transformers
python-dotenv
python-dateutil
huggingface-hub
tqdm
numpy
filelock
typing_extensions
pydantic
sse-starlette
requests
PyJWT
# llama-cpp-python ставится из wheel, собранного в Dockerfile.api
redis
httpx
pytest
pytest-asyncio
tantivy
